Enable WSDL schema validation for Standard Edition
For some reason "Use WSDL Schema Validation" checkbox was not present in TESB-SE. This is purely CXF feature and it does not require any components from TESB-EE such as Service Registry. One can just add "<entry key="schema-validation-enabled" value="true" />" to the service blueprint to achieve exactly same effect.
Lack of this switch in the UI is just pain in the a**, and it unnecessarily makes people lives harder.