fix(TESB-21796): Enforce synchronized build type flags on route build.
Currently, we have BUILD_TYPE and "camelProcessImpl.isExportMicroService()" controlling route KAR vs. microservice build. If the build type is changed in Studio in the "deployment composite", the CamelProcessImpl does not get updated, and there is no easy way to get this ensured. Therefore, the present fix ensures this update during route build.