[TESB-23646] Bug in cREST producer: response inputstream is closed (#295)
* [TESB-23646] Bug in cREST producer: response inputstream is closed

Because unmarshal() method is present in the route, after unmarshall() call InputStream is closed by Camel. But, Camel does not close InputStream if if is wrapped by Iterator. The idea of the fix is to wrap InputStream with iterator, and later unwrap it back into InputStream.

* [TESB-23646] Bug in cREST producer: response inputstream is closed

Because unmarshal() method is present in the route, after unmarshall() call InputStream is closed by Camel. But, Camel does not close InputStream if if is wrapped by Iterator. The idea of the fix is to wrap InputStream with iterator, and later unwrap it back into InputStream.

* TESB-23646 Enhance fix for variants of InputStream declaration.